requirements.txt
ase
numpy
pyyaml
scipy
matplotlib
